Change return value of withdraw method

+++ src/main/java/ua/korzh/testTask/accountService/AccountService.java
@@ -3,9 +3,9 @@ package ua.korzh.testTask.accountService;
 import ua.korzh.testTask.model.Account;
 
 public interface AccountService {
-    public Account create(double balance);
+    public Account create(long balance);
 
-    public boolean addMoney(Account account, double balance);
+    public boolean addMoney(Account account, long balance);
 
-    public double withDrawMoney(Account account, double sum);
+    public long withDrawMoney(Account account, long sum);
 }
