+++ src/main/java/ua/korzh/testTask/clientService/DepositeService.java
@@ -3,5 +3,5 @@ package ua.korzh.testTask.clientService;
 import ua.korzh.testTask.model.Client;
 
 public interface DepositeService {
-    public boolean deposite(Client client, double money);
+    public boolean deposite(Client client, long money);
 }
