+++ src/main/java/ua/korzh/testTask/clientService/WithdrawService.java
@@ -3,5 +3,5 @@ package ua.korzh.testTask.clientService;
 import ua.korzh.testTask.model.Client;
 
 public interface WithdrawService {
-    public double withdraw(Client client, double sum);
+    public long withdraw(Client client, long sum);
 }
